git & gitlab support, not working yet

+++ apps/blobserver2/blobserver2.py
@@ -12,7 +12,7 @@ import JumpScale.grid.zdaemon
 j.logger.consoleloglevel = 2
 
 import inspect
-
+import ujson
 
 class BlobserverCMDS():
 
@@ -28,11 +28,23 @@ class BlobserverCMDS():
 
         # j.logger.setLogTargetLogForwarder()
 
-    def set(self,key,value,repoId=0,session=None):
+    def set(self,key,value,repoId="",session=None):
+        if key<>"" and self.exists(key):
+            return key
         key=j.tools.hash.md5_string(value)
         backuppath=j.system.fs.joinPaths(self.STORpath,key[0:2],key[2:4],key)
         j.system.fs.createDir(j.system.fs.getDirName(backuppath))
         j.system.fs.writeFile(backuppath,value)
+        mdpath=backuppath+".md"
+        if not j.system.fs.exists(path=mdpath):
+            md={}
+        else:
+            md=ujson.loads(j.system.fs.fileGetContents(mdpath))
+        if not md.has_key("repos"):
+            md["repos"]={}
+        md["repos"][str(repoId)]=True
+        mddata=ujson.dumps(md)        
+        j.system.fs.writeFile(backuppath+".md",mddata)
         return [key,True,True]
 
     def get(self,key,session=None):
@@ -42,16 +54,24 @@ class BlobserverCMDS():
             fp.close()
         return data2
 
-    def delete(self,key,repoId=0,session=None):
-        from IPython import embed
-        print "DEBUG NOW delete"
-        embed()
-        
+    def delete(self,key,repoId="",session=None):
+        if key<>"" and not self.exists(key):
+            return 
         backuppath=j.system.fs.joinPaths(self.STORpath,key[0:2],key[2:4],key)
-        with open(backuppath) as fp:
-            data2 = fp.read()
-            fp.close()
-        return data2
+        mdpath=backuppath+".md"
+        if not j.system.fs.exists(path=mdpath):       
+            raise RuntimeError("did not find metadata")
+        md=ujson.loads(j.system.fs.fileGetContents(mdpath))
+        if not md.has_key("repos"):
+            raise RuntimeError("error in metadata on path:%s, needs to have repos as key."%mdpath)
+        if md["repos"].has_key(str(repoId)):
+            md["repos"].pop(str(repoId))
+        if md["repos"]=={}:
+            j.system.fs.remove(backuppath)
+            j.system.fs.remove(mdpath)
+        else:
+            mddata=ujson.dumps(md)        
+            j.system.fs.writeFile(backuppath+".md",mddata)
 
     def exists(self,key,session=None):
         backuppath=j.system.fs.joinPaths(self.STORpath,key[0:2],key[2:4],key)
